+++ cartography/intel/aws/inspector.py
@@ -13,6 +13,7 @@ import neo4j
 from cartography.client.core.tx import load
 from cartography.client.core.tx import load_matchlinks
 from cartography.graph.job import GraphJob
+from cartography.helpers import batch
 from cartography.intel.aws.util.botocore_config import create_boto3_client
 from cartography.models.aws.inspector.findings import AWSInspectorFindingSchema
 from cartography.models.aws.inspector.findings import InspectorFindingToPackageMatchLink
@@ -20,7 +21,6 @@ from cartography.models.aws.inspector.packages import AWSInspectorPackageSchema
 from cartography.util import aws_handle_regions
 from cartography.util import aws_paginate
 from cartography.util import AWS_REGION_ACCESS_DENIED_ERROR_CODES
-from cartography.util import batch
 from cartography.util import is_service_control_policy_explicit_deny
 from cartography.util import timeit
